api-builder: consolidate readTypescriptProcessor configuration
